Deprecate psuHead and pushTail in Arrays lib
