+++ Tests/test-deprecated.php
@@ -3,6 +3,7 @@
 declare(strict_types=1);
 
 use PHPUnit\Framework\TestCase;
+use PinkCrab\FunctionConstructors\Arrays as Arr;
 use PinkCrab\FunctionConstructors\Strings as Str;
 use PinkCrab\FunctionConstructors\FunctionsLoader;
 
@@ -163,5 +164,65 @@ class DeprecatedTest extends TestCase
     {
         $compareTheBaseAsChars = @Str\similarAsBase('BASE');
         $this->assertEquals(4, $compareTheBaseAsChars('THE BASE'));
+    }
+
+    /** @testdox Calling pushHead should throw a deprecation notice */
+    public function testPushHead(): void
+    {
+        // If using PHPUnit 9, we need to use the expectDeprecation() method
+        if (version_compare(\PHPUnit\Runner\Version::id(), '9.0.0', '>=')) {
+            $this->expectDeprecation();
+        } else {
+            $this->expectException(\PHPUnit\Framework\Error\Deprecated::class);
+        }
+        $pushToHead = Arr\pushHead(array(3, 4, 5, 6));
+        $added2     = $pushToHead(2);
+        $this->assertEquals(2, $added2[0]);
+    }
+
+
+    /** @testdox Calling pushTail should throw a deprecation notice */
+    public function testPushTail(): void
+    {
+        // If using PHPUnit 9, we need to use the expectDeprecation() method
+        if (version_compare(\PHPUnit\Runner\Version::id(), '9.0.0', '>=')) {
+            $this->expectDeprecation();
+        } else {
+            $this->expectException(\PHPUnit\Framework\Error\Deprecated::class);
+        }
+        $pushToHead = Arr\pushTail(array(3, 4, 5, 6));
+        $added2     = $pushToHead(7);
+        $this->assertEquals(7, $added2[4]);
+    }
+
+
+    public function testCanPushToHead(): void
+    {
+        $pushToHead = @Arr\pushHead(array(3, 4, 5, 6));
+        $added2     = $pushToHead(2);
+        $this->assertEquals(2, $added2[0]);
+
+        $pushToHead = @Arr\pushHead($added2);
+        $added1     = $pushToHead(1);
+        $this->assertEquals(1, $added1[0]);
+
+        // As curried.
+        $curried = @Arr\pushHead(array(3, 4, 5, 6))(2);
+        $this->assertEquals(2, $curried[0]);
+    }
+
+    public function testCanPushToTail(): void
+    {
+        $pushToTail = @Arr\pushTail(array(1, 2, 3, 4));
+        $added2     = $pushToTail(5);
+        $this->assertEquals(5, $added2[4]);
+
+        $pushToTail = @Arr\pushTail($added2);
+        $added1     = $pushToTail(6);
+        $this->assertEquals(6, $added1[5]);
+
+        // As curried.
+        $curried = @Arr\pushTail(array(1, 2, 3, 4))(5);
+        $this->assertEquals(5, $curried[4]);
     }
 }
