+++ src/arrays.php
@@ -75,9 +75,12 @@ function prepend($value): Closure
  *
  * @param array<int|string, mixed> $array
  * @return Closure(mixed):array<int|string, mixed>
+ * @deprecated 0.3.0 Use prepend() instead.
  */
 function pushHead(array $array): Closure
 {
+
+    trigger_error('Deprecated function called. This function will be removed in later versions.', E_USER_DEPRECATED);
     /**
      * @param mixed $value Adds value start of array.
      * @return array New array with value on head.
@@ -93,9 +96,11 @@ function pushHead(array $array): Closure
  *
  * @param array<int|string, mixed> $array
  * @return Closure(mixed):array<int|string, mixed>
+ * @deprecated 0.3.0 Use append() instead.
  */
 function pushTail(array $array): Closure
 {
+    trigger_error('Deprecated function called. This function will be removed in later versions.', E_USER_DEPRECATED);
     /**
      * @param mixed $value Adds value end of array.
      * @return array<int|string, mixed> New array with value on tail.
